Moved outrendering logic from modelAPI.h and replaced raw pointers with reference counting.
